refactoring based on coderabbit feedback Assisted-by: Claude Sonnet 4.5 (via Claude Code)
